pyFeedGMaps/requirements.txt
gtfs_kit==5.1.4
chardet==4.0.0
pandas==1.3.4
schedule==1.1.0
typing-extensions==3.10.0.0
wget==3.2
Unidecode==1.1.2
python-dateutil==2.8.2
